test: factor out contract-setup
